pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cursorhub"
version = "0.1.0"
description = "A macOS menu bar app for managing Cursor IDE projects, chat history, and workspaces."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Phil", email = "phil@example.com"},
]
keywords = ["cursor", "ide", "menu-bar", "macos", "developer-tools", "workspace", "productivity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "rumps>=0.4.0",
    "pyobjc-framework-Cocoa>=9.0",
]

[project.optional-dependencies]
ai = ["google-genai"]

[project.urls]
Homepage = "https://github.com/philpersonal/cursorhub"
Repository = "https://github.com/philpersonal/cursorhub"
Issues = "https://github.com/philpersonal/cursorhub/issues"

[project.scripts]
cursorhub = "cursorhub.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cursorhub = ["resources/*.png"]
